moved QueryInterfaceMetric
